+++ tests/test_pipe/test_parsing_g1830.py
@@ -213,14 +213,14 @@ class TestTranscriptPostProcessor1830(unittest.TestCase):
 
     def test_shape(self):
         self.assertEqual(1346, self.df.shape[0])
-        self.assertEqual(22, self.df.shape[1])
+        self.assertEqual(23, self.df.shape[1])
 
     def test_columns(self):
         expected = [
             'amount', 'company', 'direction', 'id', 'location', 'new_train',
             'old_train', 'parent', 'per_share', 'percentage', 'phase', 'player',
             'private', 'rotation', 'route', 'sequence', 'share_price', 'source',
-            'tile', 'train', 'type', 'result'
+            'tile', 'train', 'type', 'result', 'major_round'
         ]
         self.assertEqual(sorted(expected), sorted(list(self.df.columns)))
 
@@ -274,6 +274,14 @@ class TestTranscriptPostProcessor1830(unittest.TestCase):
         self.assertEqual('NYC', self.df.iloc[950, :].company)
         self.assertEqual('B&M', self.df.iloc[1089, :].company)
 
+    def test_major_rounds(self):
+        expected = {
+            'ISR 1',
+            'OR 1', 'OR 2', 'OR 3', 'OR 4', 'OR 5', 'OR 6', 'OR 7', 'OR 8',
+            'SR 1', 'SR 2', 'SR 3', 'SR 4', 'SR 5', 'SR 6', 'SR 7', 'SR 8'
+        }
+        self.assertEqual(expected, set(self.df.major_round.unique()))
+
 
 class TestGameStateProcessor1830(unittest.TestCase):
     df = None
@@ -305,7 +313,7 @@ class TestGameStateProcessor1830(unittest.TestCase):
 
     def test_shape(self):
         self.assertEqual(1346, self.df.shape[0])
-        self.assertEqual(166, self.df.shape[1])
+        self.assertEqual(167, self.df.shape[1])
 
     def test_final_state(self):
         self.assertIsInstance(self.final_state, dict)