+++ CHANGELOG.md
@@ -10,6 +10,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Detection for decrease of minimum bid during auction.
+
 ### Changed
 
 ### Removed